backend/requirements.txt
flask
flask_cors
pillow
flask_Session
python-dotenv
jinja2
pandas
Shapely
flask_socketio
pytest
requests
pytest-cov
flask_testing